+++ src/main/java/com/sun/wineshop/mapper/ToDtoMappers.java
@@ -2,17 +2,31 @@ package com.sun.wineshop.mapper;
 
 import com.sun.wineshop.dto.response.CategoryResponse;
 import com.sun.wineshop.dto.response.ProductResponse;
+import com.sun.wineshop.dto.response.UserResponse;
 import com.sun.wineshop.model.entity.Category;
 import com.sun.wineshop.model.entity.Product;
+import com.sun.wineshop.model.entity.User;
 
 import java.util.List;
 import java.util.stream.Collectors;
 
-public class ProductMapper {
+public class ToDtoMappers {
+
+    public static UserResponse toUserResponse(User user) {
+        return new UserResponse(
+                user.getId(),
+                user.getUsername(),
+                user.getFullName(),
+                user.getEmail(),
+                user.getPhone(),
+                user.getAddress(),
+                user.getBirthday()
+        );
+    }
 
     public static ProductResponse toProductResponse(Product product) {
         List<CategoryResponse> categoryResponses = product.getCategories().stream()
-                .map(ProductMapper::toCategoryResponse)
+                .map(ToDtoMappers::toCategoryResponse)
                 .collect(Collectors.toList());
 
         return new ProductResponse(
@@ -31,7 +45,7 @@ public class ProductMapper {
         );
     }
 
-    private static CategoryResponse toCategoryResponse(Category category) {
+    public static CategoryResponse toCategoryResponse(Category category) {
         return new CategoryResponse(
                 category.getId(),
                 category.getName(),
